Reorganizar estrutura de pastas para melhorar a intuitividade
Neste commit, foi realizada uma reorganização na estrutura de pastas do projeto para tornar mais intuitivas as áreas relacionadas à conexão com o backend e aos serviços de autenticação. Essa mudança visa facilitar a localização e compreensão do código relacionado a essas funcionalidades, melhorando a manutenção e o desenvolvimento do sistema.

- Criadas pastas específicas para conexão com o backend e serviços de autenticação
- Movidos arquivos relacionados a essas áreas para suas respectivas pastas
- Atualizados caminhos de importação nos arquivos conforme a nova estrutura de pastas
- Realizados testes para garantir a correta organização e funcionamento do código após a reorganização

Resolves: #8
