+++ src/app/auth/models/loginData.ts
@@ -1,0 +1,5 @@
+export interface DataLogin {
+    email: string,
+    password: string
+  }
+
