+++ src/app/auth/models/signupData.ts
@@ -1,0 +1,8 @@
+export interface DataUser {
+  name: string,
+  email: string,
+  cpf: string,
+  password: string,
+  repeatPassword: string,
+  agreeTerms: boolean
+}
