+++ src/app/auth/services/auth.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@angular/core';
+import { ApiService } from '../../core/services/api.service';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthService {
+
+  constructor(
+    private api: ApiService,
+    ) {}
+}
