--- src/app/data.ts
@@ -1,8 +1,0 @@
-export interface DataUser {
-  name: string,
-      email: string,
-      cpf: string,
-      password: string,
-      repeatPassword: string,
-      agreeTerms: boolean
-}
